+++ tests/Nodejs/11_Transports.test.ts
@@ -1,5 +1,5 @@
 import "ts-jest";
-import { Logger } from "../../src/index.node";
+import { Logger } from "../../src";
 
 describe("Transports", () => {
   test("attach one transport", (): void => {
