+++ tests/Nodejs/15_Placeholders.test.ts
@@ -1,5 +1,5 @@
 import "ts-jest";
-import { IMeta, Logger } from "../../src/index.node";
+import { IMeta, Logger } from "../../src";
 import { getConsoleLog, mockConsoleLog } from "./helper.js";
 
 describe("Placeholders", () => {
