+++ tests/Nodejs/3_hidden_loglevel.test.ts
@@ -1,5 +1,5 @@
 import "ts-jest";
-import { Logger } from "../../src/index.node";
+import { Logger } from "../../src";
 import { getConsoleLog, mockConsoleLog } from "./helper.js";
 
 const logger = new Logger({ type: "hidden" });
