+++ tests/Nodejs/9_pretty_Styles.test.ts
@@ -1,5 +1,5 @@
 import "ts-jest";
-import { Logger } from "../../src/index.node";
+import { Logger } from "../../src";
 import { getConsoleLog, mockConsoleLog } from "./helper.js";
 
 describe("Pretty: Styles", () => {
